🔨 Feat | 게틀링 코드 수정

+++ Assets/02.Scripts/Monster/MudMonster.cs
@@ -25,7 +25,7 @@ public class MudMonster : Monster
         }
     }
 
-    public float GetHealth()
+    public int GetHealth()
     {
         return Health;
     }
